Set up minimal CI on Windows and Ubuntu (#1)
This PR sets up basic CI for an empty hello-world project. It runs on both Windows and Linux using MSVC and Clang++ compilers. Tooling is deliberately minimal for now to minimize dependencies.

This commit shows that the build step fails when the program fails to compile: c6b2b7f

This commit shows that the test step fails when the tests fail: b21a88d