Write test output to injected stream instead of std::cerr (#4)
Assertions were writing directly to std::cerr. This makes it impossible to test the output of the assertions, and also garbles the unit tests of the test framework.

This PR passes a TestImpl_Stream parameter to Test::Run(). The TestImpl_ prefix is meant to reduce conflicts. ASSERT uses this hidden parameter to write its information.